feat: deploy (#13)
This feature introduces two workflows; deploy and test. Test is currently only configured to work for Elixir since we do not have any React tests as of yet. The most important addition is the deploy workflow.

Deploy interacts with Docker to create a Docker image with both our Elixir game server, API and web client. The API in this case serves our web client as well as servers the favor endpoint. This in turn is deployed on Digital Ocean utilizing docker-compose to start a nginx server with Let's Encrypt certificates and our own image which is pushed to Github registry.

Co-authored-by: jvhoven <contact@jvhoven.nl>