fix(setup): user identification
This fix adds a user property which can
be used to identify the "current" playing
user.

Solves the issue where the UI always built
logic upon the first user, that created
the lobby. Resulting in when the first user
clicked on "ready", it cascaded to the other
as well and it started the game.